fixed double.toUInt

+++ src/main/scala/is/hail/utils/richUtils/RichDouble.scala
@@ -3,5 +3,5 @@ package is.hail.utils.richUtils
 import is.hail.utils.UInt
 
 class RichDouble(val d: Double) extends AnyVal {
-  def toUInt: UInt = UInt(d.toInt)
+  def toUInt: UInt = UInt(d.toLong)
 }
